fixed: Project Structure
