+++ app/src/main/java/com/example/myrecipeapp/navigation/AppScreens.kt
@@ -1,4 +1,4 @@
-package com.example.myrecipeapp
+package com.example.myrecipeapp.navigation
 
 sealed class Screen(val route: String) {
     object Categories : Screen("categories_screen")
