+++ app/src/main/java/com/example/myrecipeapp/ui/features/mealsbycategory/MealsByCategoryScreen.kt
@@ -1,4 +1,4 @@
-package com.example.myrecipeapp
+package com.example.myrecipeapp.ui.features.mealsbycategory
 
 import android.util.Log
 import androidx.compose.foundation.Image
@@ -31,6 +31,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import coil.compose.rememberAsyncImagePainter
+import com.example.myrecipeapp.data.MealDetails
 
 @Composable
 fun CategoryDetailScreen(strCategory: String?, navController: NavController)
